\name{NetworkMaker}
\docType{class}

% the class
\alias{NetworkMaker}
\alias{class:NetworkMaker}
\alias{NetworkMaker-class}

\alias{getPackage,NetworkMaker-method}
\alias{getPackage}
\alias{\S4method{getPackage}{NetworkMaker}}

\title{NetworkMaker: Create sample/gene cyjs networks}

\description{
   to be written...
}


\usage{
NetworkMaker(packageName, verbose=FALSE)

% class methods
\S4method{getPackage}{NetworkMaker}(obj)
}

\arguments{
  \item{obj}{The \code{NetworkMaker} object returned by the
     class constructor.}

  \item{packageName}{The name of an existing and installed
    SttrDataPackage}
}

\author{Paul Shannon}

\keyword{classes}
\keyword{methods}


